refactor: move JSON creation to base class
